+++ PW/atomic_wfc.f90
@@ -84,7 +84,7 @@ subroutine atomic_wfc (ik, wfcatom)
               ux = 1.d0 - px
               vx = 2.d0 - px
               wx = 3.d0 - px
-              i0 = qg (ig) / dq + 1
+              i0 = INT( qg (ig) / dq ) + 1
               i1 = i0 + 1
               i2 = i0 + 2
               i3 = i0 + 3
